Work so far.

+++ Tuna/tunalib/initializer_list.h
@@ -46,7 +46,7 @@ namespace Tuna
       typedef _E 		value_type;
       typedef const _E& 	reference;
       typedef const _E& 	const_reference;
-      typedef size_t 		size_type;
+      typedef uint8 		size_type;
       typedef const _E* 	iterator;
       typedef const _E* 	const_iterator;
